+++ tests/test_hessian.py
@@ -4,15 +4,20 @@ import numpy as np
 import autode as ade
 from . import testutils
 from autode.atoms import Atom, Atoms
-from autode.methods import ORCA
+from autode.methods import ORCA, XTB
 from autode.calculation import Calculation
 from autode.species import Molecule
 from autode.values import Frequency
-from autode.hessians import Hessian
 from autode.geom import calc_rmsd
 from autode.units import wavenumber
 from autode.exceptions import CalculationException
 from autode.transition_states.base import displaced_species_along_mode
+from autode.values import Distance
+from autode.wrappers.keywords import HessianKeywords, GradientKeywords
+from autode.hessians import (Hessian,
+                             calculate_numerical_hessian,
+                             _NumericalHessianCalculator)
+
 here = os.path.dirname(os.path.abspath(__file__))
 
 
@@ -376,3 +381,107 @@ def test_extract_wrong_molecule_hessian():
     # N atoms (1 here)
     with pytest.raises(CalculationException):
         _ = calc.get_hessian()
+
+
+def test_num_hess_invalid_input():
+
+    water = Molecule(smiles='O')
+    orca = ORCA()
+
+    # Keywords must be GradientKeywords that don't include any 'Hessian'
+    # or 'frequency' keywords
+    for invalid_kwds in (None,
+                         GradientKeywords(['Freq', 'PBE', 'Def2-SVP']),
+                         HessianKeywords(['PBE', 'Def2-SVP'])):
+
+        with pytest.raises(ValueError):
+            calculate_numerical_hessian(species=water,
+                                        method=orca,
+                                        keywords=invalid_kwds,
+                                        do_c_diff=False,
+                                        shift=Distance(1E-3, units='Å'))
+
+
+@testutils.work_in_zipped_dir(os.path.join(here, 'data', 'num_hess.zip'))
+def test_h2_hessian():
+
+    h2 = Molecule(name='H2', atoms=[Atom('H'), Atom('H', x=0.77)])
+
+    h2.calc_hessian(method=ORCA(), numerical=False)
+    analytic_hessian = h2.hessian.copy()
+
+    h2.hessian = None  # Clear the analytic Hessian
+
+    h2.calc_hessian(method=ORCA(), numerical=True)
+    assert h2.hessian is not None
+
+    num_hessian = h2.hessian.copy()
+
+    assert np.allclose(analytic_hessian,
+                       num_hessian,
+                       atol=1E-2)
+
+
+@testutils.work_in_zipped_dir(os.path.join(here, 'data', 'num_hess.zip'))
+def test_h2_c_diff_hessian():
+
+    h2 = Molecule(name='H2', atoms=[Atom('H'), Atom('H', x=0.77)])
+
+    h2.calc_hessian(method=ORCA(), numerical=False)
+    analytic_hessian = h2.hessian.copy()
+
+    h2.hessian = None  # Clear the analytic Hessian and calculate a numerical
+    h2.calc_hessian(method=ORCA(), numerical=True, use_central_differences=True)
+
+    # Central differences should afford a very good Hessian cf. analytic
+    assert np.allclose(analytic_hessian,
+                       h2.hessian,
+                       atol=1E-3)
+
+
+@testutils.work_in_zipped_dir(os.path.join(here, 'data', 'num_hess.zip'))
+@testutils.requires_with_working_xtb_install
+def test_h2_xtb_vs_orca_hessian():
+
+    h2 = Molecule(name='H2', atoms=[Atom('H'), Atom('H', x=0.77)])
+
+    h2.calc_hessian(method=ORCA(), numerical=False)
+    orca_hessian = h2.hessian.copy()
+
+    h2.calc_hessian(method=XTB(), numerical=True, use_central_differences=True)
+    xtb_hessian = h2.hessian.copy()
+
+    # ORCA and XTB Hessians should be similar, ish
+    assert np.allclose(orca_hessian,
+                       xtb_hessian,
+                       atol=0.3)
+
+
+@testutils.work_in_zipped_dir(os.path.join(here, 'data', 'num_hess.zip'))
+@testutils.requires_with_working_xtb_install
+def test_ind_num_hess_row():
+    """Calculate d^2E/dx0dx0 using numerical displacements with and
+    without central differences"""
+
+    h2 = Molecule(name='H2', atoms=[Atom('H'), Atom('H', x=0.77)])
+    xtb = XTB()
+
+    for flag in (True, False):
+
+        calculator = _NumericalHessianCalculator(species=h2,
+                                                 method=xtb,
+                                                 keywords=xtb.keywords.grad,
+                                                 do_c_diff=flag,
+                                                 shift=0.001)
+
+        # Non central differences require an initial gradient at the curr geom
+        calculator._init_gradient = calculator._gradient(calculator._species)
+
+        if flag:
+            row = calculator._cdiff_row(atom_idx=0, component=0)
+        else:
+            row = calculator._diff_row(atom_idx=0, component=0)
+
+        assert np.isclose(row[0],
+                          1.00957,
+                          atol=1E-1)